brain/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0
pyyaml>=6.0
schedule>=1.2.0
trafilatura>=1.6.0
